+++ src/drivers/process/index.js
@@ -1,38 +1,41 @@
 const driverCore = require('../../driver-core');
-const get = require('lodash/get');
 
 const DRIVER_NAME = 'process';
 
 const processOperations = global.afDriverProcessStore = {};
 
-class processDriver extends driverCore {
-  constructor (data) {
+class ProcessDriver extends driverCore {
+  constructor({
+    clientId,
+    descriptionHash,
+  } = {}) {
     super({
-      driver: DRIVER_NAME
+      driver: DRIVER_NAME,
     });
 
-    Object.assign(this, {
-      clientId: data.clientId,
-      descriptionHash: data.descriptionHash
-    });
+    this.clientId = clientId;
+    this.descriptionHash = descriptionHash;
 
     if (!processOperations[this.descriptionHash]) {
       processOperations[this.descriptionHash] = [];
     }
   }
 
-  async join () {
+  async join() {
     this.leave();
     processOperations[this.descriptionHash].push(this.clientId);
   }
 
-  async isFirst () {
+  async isFirst() {
     return processOperations[this.descriptionHash][0] === this.clientId;
   }
 
-  async leave () {
-    processOperations[this.descriptionHash] = processOperations[this.descriptionHash].filter((cid) => this.clientId !== cid);
+  async leave() {
+    processOperations[this.descriptionHash] = processOperations[this.descriptionHash]
+      .filter(
+        cid => this.clientId !== cid,
+      );
   }
 }
 
-module.exports = processDriver;
+module.exports = ProcessDriver;