+++ test/test.js
@@ -1,16 +1,33 @@
 const test = require('ava');
 const conf = require('./mongo-conf');
 
-const AF = require('../')(conf);
+const sessionName = new Date().getTime();
+
+const AF_MONGO = require('../')({
+  ...conf,
+  sessionName,
+  driverName: 'mongodb',
+  noSHA: true,
+});
 const AF_PROCESS = require('../')({
-  driverName: 'process'
+  driverName: 'process',
+  awaitTimeoutSec: 900,
+  noSHA: true,
+});
+
+const AF_REDIS = require('../')({
+  driverName: 'redis',
+  sessionName: `afredis.${sessionName}`,
+  noSHA: true,
 });
 
 const someCustomClass = require('../src/drivers/mongodb');
 
 const AF_CUSTOM = require('../')({
   driverName: 'custom',
-  driverClass: someCustomClass
+  driverClass: someCustomClass,
+  sessionName,
+  noSHA: true,
 });
 
 async function sleep (ms) {
@@ -64,19 +81,19 @@ const FLOWS = [
 // tests
 
 test('(mongodb) one action flow', async (t) => {
-  const af = () => AF.create(FLOWS[0]);
+  const af = () => AF_MONGO.create(FLOWS[0]);
 
   await actionFlow(af, t);
 });
 
 test('(mongodb) multi action flow (all blocks)', async (t) => {
-  const af = () => AF.multi(FLOWS);
+  const af = () => AF_MONGO.multi(FLOWS);
 
   await actionFlow(af, t);
 });
 
 test('(mongodb) multi action flow (last 2 blocks)', async (t) => {
-  const af = () => AF.multi([
+  const af = () => AF_MONGO.multi([
     FLOWS[2], FLOWS[3]
   ]);
 
@@ -84,7 +101,7 @@ test('(mongodb) multi action flow (last 2 blocks)', async (t) => {
 });
 
 test('(mongodb) multi action flow (first 2 blocks)', async (t) => {
-  const af = () => AF.multi([
+  const af = () => AF_MONGO.multi([
     FLOWS[0], FLOWS[1]
   ]);
 
@@ -92,7 +109,7 @@ test('(mongodb) multi action flow (first 2 blocks)', async (t) => {
 });
 
 test('(mongodb) multi action flow (last 3 blocks)', async (t) => {
-  const af = () => AF.multi([
+  const af = () => AF_MONGO.multi([
    FLOWS[1],
    FLOWS[2],
    FLOWS[3]
@@ -115,6 +132,18 @@ test('(process) multi action flow', async (t) => {
 
 test('(custom) multi action flow', async (t) => {
   const af = () => AF_CUSTOM.multi(FLOWS);
+
+  await actionFlow(af, t);
+});
+
+test('(redis) one action flow', async (t) => {
+  const af = () => AF_REDIS.create(FLOWS[0]);
+
+  await actionFlow(af, t);
+});
+
+test('(redis) multi action flow', async (t) => {
+  const af = () => AF_REDIS.multi(FLOWS);
 
   await actionFlow(af, t);
 });