+++ CHANGELOG.md
@@ -1,10 +1,15 @@
 # Change log
 
+## v2.2
+- **2.2.0**
+  - Code refactoring
+  - Update docs
+  - Add redis driver, and set by default
 ## v2.1
 - **2.1.0**
   - Optimize code.
   - Fix bugs of deadlocks at multi mode.
   - Add tests.
-  - Refactory: Move the storage logic to drivers. 
+  - Refactoring: Move the storage logic to drivers. 
   - Added possibility create custom drivers.
   - Added the Process driver. For more see docs.